requirements.txt
nicegui
requests
pillow
qrcode
pywin32; sys_platform == "win32"
pycups; sys_platform != "win32"
pytest
python-dotenv
